Working on the UI for the scoring screen
I realized that, at least for the app's first version, I only want to score at the end of the game and not on each stage (This is actually how scoring is done when playing the game). Therefore, I only need one screen for scoring, instead of one for each stage.