Fixed console error "Cannot update a component (`App`) while rendering a different component"
I was getting this when trying to navigate to the Game result page inside the getGameResult function; I was calling this function inside a nested turnary operator.

According to the accepted reply here https://stackoverflow.com/questions/62236000/cannot-update-a-component-app-while-rendering-a-different-component, the error message will be displayed if parent (App) states are set while rendering children.

I fixed this by adding the navigation inside a useEffect hook.